+++ docs/Changelog.md
@@ -3,6 +3,7 @@
 ## v4.3
 * Make switchFragment public
 * Create addFragment methods for backstacks
+* Set fab events to onStart
 
 ## v4.2
 * Code cleanup